fix(crabbox): satisfy catalog and warm-image lint checks
Preserve operating-system catalog output without map spreads. Move warm-image timeout policy and fork response validation into the existing timeout and checkpoint modules to keep the manager within its line budget.
